+++ opaeum-runtime/opaeum-runtime-domain/src/main/java/org/opaeum/runtime/environment/marshall/PropertyValue.java
@@ -4,13 +4,13 @@ import java.io.Serializable;
 
 public class PropertyValue implements Serializable{
 	private static final long serialVersionUID = -5852805763067486690L;
-	private Integer id;
+	private Long id;
 	private Value value;
-	public PropertyValue(Integer id,Value value){
+	public PropertyValue(Long id,Value value){
 		this.id = id;
 		this.value=value;
 	}
-	public Integer getId(){
+	public Long getId(){
 		return id;
 	}
 	public Value getValue(){
